requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
cloudinary==1.42.2
Django==5.1.6
django-ckeditor-5==0.2.13
django-environ==0.12.0
django-filter==25.1
django-js-asset==3.1.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
idna==3.10
Markdown==3.7
pillow==11.1.0
PyJWT==2.9.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
urllib3==2.3.0
